+++ c2pa_c_ffi/src/lib.rs
@@ -20,12 +20,14 @@ mod error;
 mod json_api;
 mod signer_info;
 
-pub use c2pa::{AsyncSigner, Builder, Reader, Signer, SigningAlg};
+pub use c2pa::{
+    AsyncSigner, Builder, Error as C2paError, Reader, Result as C2paResult, Signer, SigningAlg,
+};
 pub use c2pa_stream::*;
 pub use c_api::*;
 // Re-export for macro use
 #[doc(hidden)]
 pub use cimpl::cimpl_error::CimplError;
 pub use cimpl::*;
-pub use error::{C2paError, Error, Result};
+pub use error::{Error, Result};
 pub use signer_info::SignerInfo;